+++ camera-ui-openvino/src/main.py
@@ -770,6 +770,10 @@ class OpenVinoPlugin(
             return_exceptions=True,
         )
 
+        for sensors in self._sensors.values():
+            for sensor in sensors.values():
+                sensor.updateModelSpec()
+
     async def _reset_settings(self) -> None:
         await reset_stored_settings(self.storage)
         self.logger.log("Settings reset to defaults")